get_trips - define the time frame!
Start / stop dates are strings in ISO 8601 format. 
If no timezone is given, the local timezone is used if available. If not, UTC is assumed. The dates + times are then converted to UTC for simplicity.
CAVE: The search on the JLR servers is recent-to-old.

If no start and/or stop dates are given, the behaviour of get_trips is unchanged wrt the function it replaces.